change return type from size_t to int
git-svn-id: http://svn.statmt.org/repository/mira@3751 cc96ff50-19ce-11e0-b349-13d7f0bd23df

+++ branches/mira-mtm5/mira/Main.cpp
@@ -378,7 +378,7 @@ int main(int argc, char** argv) {
 		  // run optimiser on batch
 	      cerr << "\nRun optimiser.." << endl;
 	      ScoreComponentCollection oldWeights(mosesWeights);
-	      size_t constraintChange = optimiser->updateWeights(mosesWeights, featureValues, losses, oracleFeatureValues);
+	      int constraintChange = optimiser->updateWeights(mosesWeights, featureValues, losses, oracleFeatureValues);
 
 		  // update moses weights
 	      mosesWeights.L1Normalise();
