+++ branches/mira-mtm5/mira/Optimiser.h
@@ -29,7 +29,7 @@ namespace Mira {
   class Optimiser {
     public:
       Optimiser() {}
-      virtual size_t updateWeights(Moses::ScoreComponentCollection& weights,
+      virtual int updateWeights(Moses::ScoreComponentCollection& weights,
                          const std::vector<std::vector<Moses::ScoreComponentCollection> >& scores,
                          const std::vector<std::vector<float> >& losses,
                          const std::vector<Moses::ScoreComponentCollection>& oracleScores) = 0;
@@ -37,7 +37,7 @@ namespace Mira {
  
   class DummyOptimiser : public Optimiser {
     public:
-      virtual size_t updateWeights(Moses::ScoreComponentCollection& weights,
+      virtual int updateWeights(Moses::ScoreComponentCollection& weights,
                          const std::vector< std::vector<Moses::ScoreComponentCollection> >& scores,
                          const std::vector< std::vector<float> >& losses,
                          const std::vector<Moses::ScoreComponentCollection>& oracleScores)
@@ -48,7 +48,7 @@ namespace Mira {
     public:
        
 
-      virtual size_t updateWeights(Moses::ScoreComponentCollection& weights,
+      virtual int updateWeights(Moses::ScoreComponentCollection& weights,
                          const std::vector< std::vector<Moses::ScoreComponentCollection> >& scores,
                          const std::vector< std::vector<float> >& losses,
                          const std::vector<Moses::ScoreComponentCollection>& oracleScores);
@@ -71,7 +71,7 @@ namespace Mira {
 
      ~MiraOptimiser() {}
    
-      virtual size_t updateWeights(Moses::ScoreComponentCollection& weights,
+      virtual int updateWeights(Moses::ScoreComponentCollection& weights,
       						  const std::vector< std::vector<Moses::ScoreComponentCollection> >& scores,
       						  const std::vector< std::vector<float> >& losses,
       						  const std::vector< Moses::ScoreComponentCollection>& oracleScores);
